Reorganize `storage` crate modules
