+++ node/libs/storage/src/types.rs
@@ -1,10 +1,15 @@
 //! Defines the schema of the database.
 
 use anyhow::Context as _;
+use concurrency::ctx;
 use rocksdb::{Direction, IteratorMode};
-use roles::validator;
+use roles::validator::{self, BlockNumber};
 use schema::{proto::storage as proto, read_required, ProtoFmt};
-use std::collections::{BTreeMap, HashMap};
+use std::{
+    collections::{BTreeMap, HashMap},
+    iter, ops,
+};
+use thiserror::Error;
 
 /// Enum used to represent a key in the database. It also acts as a separator between different stores.
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -14,7 +19,7 @@ pub(crate) enum DatabaseKey {
     ReplicaState,
     /// Key used to store the finalized blocks.
     /// Block(BlockNumber) -> FinalBlock
-    Block(validator::BlockNumber),
+    Block(BlockNumber),
 }
 
 impl DatabaseKey {
@@ -43,11 +48,11 @@ impl DatabaseKey {
     }
 
     /// Parses the specified bytes as a `Self::Block(_)` key.
-    pub(crate) fn parse_block_key(raw_key: &[u8]) -> anyhow::Result<validator::BlockNumber> {
+    pub(crate) fn parse_block_key(raw_key: &[u8]) -> anyhow::Result<BlockNumber> {
         let raw_key = raw_key
             .try_into()
             .context("Invalid encoding for block key")?;
-        Ok(validator::BlockNumber(u64::from_be_bytes(raw_key)))
+        Ok(BlockNumber(u64::from_be_bytes(raw_key)))
     }
 }
 
@@ -64,7 +69,7 @@ pub struct ReplicaState {
     pub high_qc: validator::CommitQC,
     /// A cache of the received block proposals.
     pub block_proposal_cache:
-        BTreeMap<validator::BlockNumber, HashMap<validator::BlockHash, validator::Block>>,
+        BTreeMap<BlockNumber, HashMap<validator::BlockHash, validator::Block>>,
 }
 
 impl ProtoFmt for ReplicaState {
@@ -115,3 +120,68 @@ impl ProtoFmt for ReplicaState {
         }
     }
 }
+
+/// Iterator over missing block numbers.
+pub(crate) struct MissingBlockNumbers<I: Iterator> {
+    range: ops::Range<BlockNumber>,
+    existing_numbers: iter::Peekable<I>,
+}
+
+impl<I> MissingBlockNumbers<I>
+where
+    I: Iterator<Item = anyhow::Result<BlockNumber>>,
+{
+    /// Creates a new iterator based on the provided params.
+    pub(crate) fn new(range: ops::Range<BlockNumber>, existing_numbers: I) -> Self {
+        Self {
+            range,
+            existing_numbers: existing_numbers.peekable(),
+        }
+    }
+}
+
+impl<I> Iterator for MissingBlockNumbers<I>
+where
+    I: Iterator<Item = anyhow::Result<BlockNumber>>,
+{
+    type Item = anyhow::Result<BlockNumber>;
+
+    fn next(&mut self) -> Option<Self::Item> {
+        // Loop while existing numbers match the starting numbers from the range. The check
+        // that the range is non-empty is redundant given how `existing_numbers` are constructed
+        // (they are guaranteed to be lesser than the upper range bound); we add it just to be safe.
+        while !self.range.is_empty()
+            && matches!(self.existing_numbers.peek(), Some(&Ok(num)) if num == self.range.start)
+        {
+            self.range.start = self.range.start.next();
+            self.existing_numbers.next(); // Advance to the next number
+        }
+
+        if matches!(self.existing_numbers.peek(), Some(&Err(_))) {
+            let err = self.existing_numbers.next().unwrap().unwrap_err();
+            // ^ Both unwraps are safe due to the check above.
+            return Some(Err(err));
+        }
+
+        if self.range.is_empty() {
+            return None;
+        }
+        let next_number = self.range.start;
+        self.range.start = self.range.start.next();
+        Some(Ok(next_number))
+    }
+}
+
+/// Storage errors.
+#[derive(Debug, Error)]
+pub enum StorageError {
+    /// Operation was canceled by structured concurrency framework.
+    #[error("operation was canceled by structured concurrency framework")]
+    Canceled(#[from] ctx::Canceled),
+    /// Database operation failed.
+    #[error("database operation failed")]
+    Database(#[source] anyhow::Error),
+}
+
+/// [`Result`] for fallible storage operations.
+pub type StorageResult<T> = Result<T, StorageError>;
